[Solaris] Default to -fno-cxa-finalize
There is no __cxa_finalize symbol available on Solaris, so stop
looking for cxa_finalize.o and add -fno-use-cxa-atexit as
default. Years ago this was reverted (r151648) because
-fno-use-cxa-atexit was apparently broken, but it seems to work fine
now. In any case we'd need to ship an implementation to be able to use
the feature, otherwise we'll just fail to link C++ programs.